+++ test/util/async/test_stream_take.rs
@@ -4,7 +4,7 @@ use super::nums;
 #[test]
 pub fn test_stream_take() {
     let stream = nums(0, 10).take(4);
-    let vals: Vec<usize> = stream.iter().collect();
+    let vals: Vec<i32> = stream.iter().collect();
 
     assert_eq!([0, 1, 2, 3].as_slice(), vals.as_slice());
 }
